project/requirements.txt
adafruit-blinka
numpy
playsound
pygobject
pycario
RPi.GPIO
scipy
soundfile
wave